Remove router and define routes directly in app
